requirements.txt
streamlit==1.46.1
easyocr==1.7.2
opencv-python-headless==4.12.0.88
Pillow==11.3.0